engject/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
cryptography==46.0.5
cssselect==1.4.0
distro==1.9.0
Django==6.0.3
feedfinder2==0.0.4
filelock==3.25.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-auth==2.49.0
google-auth-httplib2==0.3.0
google-genai==1.66.0
google-generativeai==0.8.6
googleapis-common-protos==1.73.0
groq==1.0.0
grpcio==1.78.0
grpcio-status==1.71.2
gunicorn==25.1.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
jieba3k==0.35.1
joblib==1.5.3
lxml==6.0.2
lxml_html_clean==0.4.4
newspaper3k==0.2.8
nltk==3.9.3
packaging==26.0
pillow==12.1.1
proto-plus==1.27.1
protobuf==5.29.6
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
requests-file==3.0.1
rsa==4.9.1
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
sqlparse==0.5.5
tenacity==9.1.4
tinysegmenter==0.3
tldextract==5.3.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
websockets==16.0
whitenoise==6.12.0
pytrends==4.9.2